Rotas para post no blog

+++ src/controllers/api/UserController.php
@@ -4,25 +4,30 @@
 
 namespace Src\Controllers\Api;
 
+use Src\Controllers\Api\ResponseController;
 use Src\Core\App;
-use Src\User\User;
 use Src\Core\Request;
+use Src\User;
 
 
-class UserController 
+class UserController extends ResponseController
 {
 	public function index()
 	{
 
 		$users = User::all();
-		dd($users);
+		return $this->respond($users);
 	}
 
 	public function remove($id)
 	{
 		$user = User::find($id);
+		if (empty($user)) {
+			return $this->respondNotFound('Usuario não encontrado');
+		}
+
 		$user->delete();
-		dd($user);
+		return $this->respond();
 	}
 
 	public function store()
@@ -31,25 +36,31 @@ class UserController
 		$parameter = ['name' => Request::getParameters()['name'] ];
 
 		User::insert($parameter);
-		dd('foi');
+		return $this->respond();
 	}
 
 	public function get($id)
 	{
 
 		$user = User::find($id);
-		dd($user);
+		if (empty($user)) {
+			return $this->respondNotFound('Usuario não encontrado');
+		}
+		return $this->respond($user);
 	}
 
 	public function update($id)
 	{
 
 		$user = User::find($id);
+		if (empty($user)) {
+			return $this->respondNotFound('Usuario não encontrado');
+		}
 		$parameter = ['name' => Request::getParameters()['name'],
 					];
 
 		$user->update($parameter);
-		dd('foi');
+		return $this->respond();
 	}
